NAE Ready/requirements.txt
pyautogen==0.9.0
autogenstudio>=0.2.0
openai>=1.0.0
python-dotenv>=1.0.0
psutil>=5.9.0
requests>=2.26.0
numpy>=1.21.0
pandas>=1.3.0
yfinance>=0.2.0
matplotlib>=3.5.0
scipy>=1.7.0

# Redis for kill switch and state management
redis>=4.5.0

# Trading API clients
alpaca-trade-api>=3.0.0  # Legacy SDK (for backward compatibility)
alpaca-py>=0.26.0  # Official Alpaca Python SDK with TradingClient
ibapi>=10.0.0

# Market data APIs
polygon-api-client>=1.0.0

# Security and encryption
cryptography>=3.4.8
bcrypt>=3.2.0

# Web framework for API endpoints
fastapi>=0.100.0
uvicorn>=0.20.0

# Database
sqlalchemy>=1.4.0
alembic>=1.8.0

# Monitoring and logging
prometheus-client>=0.15.0
structlog>=22.0.0

# Configuration management
pydantic>=1.10.0
pydantic-settings>=2.0.0

# Production deployment
gunicorn>=20.1.0
supervisor>=4.2.0

# Web scraping and API integrations
beautifulsoup4>=4.11.0
selenium>=4.15.0
tweepy>=4.14.0
praw>=7.7.0
discord.py>=2.3.0
yfinance>=0.2.0
alpha-vantage>=2.3.0
newsapi-python>=0.2.6

# TD3 Stock Agent (Deep-RL-Stocks port) — enables training; numpy fallback if not installed
torch>=2.0.0

# Data processing and validation
scikit-learn>=1.3.0
textblob>=0.17.1
nltk>=3.8.0
spacy>=3.6.0

# Rate limiting and caching
ratelimit>=2.2.1
cachetools>=5.3.0

# Payment integrations
paypalcheckoutsdk>=1.0.4

# Optional secret management (AWS Secrets Manager)
boto3>=1.28.0

# OAuth for E*Trade API
requests-oauthlib>=1.3.0
oauthlib>=3.2.0

# New dependencies for enhanced NAE features
anthropic>=0.18.0  # For Claude Sonnet 4.5 models
typing-extensions>=4.8.0  # Enhanced type hints

# Casey Interface Dependencies
flask>=2.3.0
flask-cors>=4.0.0
websockets>=11.0

# THRML (Thermodynamic Machine Learning) for probabilistic models and energy-based learning
jax>=0.4.0
jaxlib>=0.4.0
# Note: THRML may need to be installed from Extropic's repository
# pip install git+https://github.com/extropic-ai/thrml.git
# Or: pip install thrml (if available on PyPI)

# Robustness & Monitoring
prometheus-client>=0.18.0
schedule>=1.2.0

# Data processing
pyarrow>=10.0.0  # For Parquet support
fastparquet>=2023.0.0  # Alternative Parquet backend

# YouTube transcript fetching
yt-dlp>=2023.12.0  # For fetching YouTube transcripts (legal for transcripts)

# =============================================================================
# POLYMARKET INTEGRATION (Prediction Markets)
# Based on official Polymarket GitHub repositories:
# - github.com/Polymarket/py-clob-client
# - github.com/Polymarket/agents
# =============================================================================

# Polymarket CLOB SDK (Central Limit Order Book)
py-clob-client>=0.16.0  # Official Polymarket Python SDK

# Web3 for Polygon blockchain interaction
web3>=6.0.0  # Ethereum/Polygon interface

# HTTP client (faster async requests)
httpx>=0.25.0  # Modern async HTTP client

# LangChain for AI-powered trading (used by official Polymarket agents)
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5  # For GPT-based superforecasting

# Pydantic for data validation (Polymarket data models)
pydantic>=2.0.0

# =============================================================================
# KALSHI INTEGRATION (CFTC-Regulated Prediction Markets)
# Based on official Kalshi resources:
# - Official kalshi-python SDK
# - Kalshi API documentation (docs.kalshi.com)
# - github.com/Kalshi/tools-and-analysis
# =============================================================================

# Kalshi Official Python SDK
kalshi-python>=1.0.0  # Official Kalshi SDK

# Cryptography for RSA-PSS authentication
cryptography>=41.0.0  # RSA key handling for Kalshi API auth

# PyYAML for credentials (optional, used by some Kalshi examples)
pyyaml>=6.0